Unify candle fetching pipeline
